refactor: update shared components folder structure
