+++ AngularStandalone/src/app/shared/components/lists/document-list-item/document-list-item.component.scss
