+++ AngularStandalone/src/app/shared/components/lists/expansion-document-list/categorized-search-result.model.ts
