+++ AngularStandalone/src/app/shared/components/lists/expansion-document-list/expansion-document-list.component.scss
