+++ AngularStandalone/src/app/shared/components/markdown/document-header/document-header.component.scss
