+++ AngularStandalone/src/app/shared/components/markdown/document-toc/document-toc.component.scss
