+++ AngularStandalone/src/app/shared/components/markdown/document/document.component.scss
