+++ AngularStandalone/src/app/shared/components/tags/tag/tag.component.scss
